Refactor exchange rate handling from Decimal to double
- Changed the `rate` field in `ExchangeRateEntry` from `Decimal` to `double` for consistency and ease of use.
- Updated related mappers, repositories, and services to accommodate the new double type for exchange rates.
- Adjusted parsing logic to ensure correct handling of exchange rates with both comma and dot decimal formats.
- Modified tests to reflect the changes in rate type and ensure proper functionality.
- Enhanced the user interface to allow selection of a primary currency and display equivalent amounts based on the new rate handling.

+++ lib/data/models/exchange_rate_entry.dart
@@ -1,19 +1,19 @@
-import 'package:decimal/decimal.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'exchange_rate_entry.freezed.dart';
 
-/// Domain model for a cached exchange-rate observation (PROJECT_PLAN §5.2).
+/// Domain model for a cached exchange-rate observation.
 ///
-/// `rate` is the exact `baseCurrency → quoteCurrency` factor. `asOfDate` is a
-/// date-only value. On create, `id` is assigned by the database and ignored.
+/// `rate` is the raw `baseCurrency -> quoteCurrency` multiplier. It is
+/// independent of currency minor digits. On create, `id` is assigned by the
+/// database and ignored.
 @freezed
 abstract class ExchangeRateEntry with _$ExchangeRateEntry {
   const factory ExchangeRateEntry({
     required int id,
     required String baseCurrency,
     required String quoteCurrency,
-    required Decimal rate,
+    required double rate,
     required DateTime asOfDate,
     String? source,
   }) = _ExchangeRateEntry;
